requirements.txt
colorzero==2.0
gpiozero==2.0.1
mpu6050-raspberrypi==1.2
smbus2==0.4.3
